Add TODO comments regarding CTRE SignalLogger in drivetrain-sysid project
The basic drivetrain-sysid project, like all the WPILib examples, doesn't support CTRE motor controllers directly so we had to port it for use with our Krakens. Reading the documentation, however, it seems that CTRE provide a high resolution motor logging class, SignalLogger, that seems to be recommended over the low resolution, general purpose functions used in the example. To use this, we will need to make several more changes to the project commands to configure, start and stop CTRE logging when running each of the test scenarios. The current code includes command factories and I don't yet know how to modify the syntax to create new versions of these with our additions. I must learn Java...