Changed motor mode from coast to brake globally
When we e-stopped, the robot would coast and potentially hit things, we changed it to a hard brake so when we e-stop the robot stops immediately instead of coasting. We applied this globally so when we stop while driving (completely let go of the joystick) the robot hard brakes. This commit also halves the max drive speed for demonstration purposes.